+++ animator/src/cs3500/animator/util/myUtil.java
@@ -11,15 +11,15 @@ public class MyUtil {
    * @param time
    * @return
    */
-  public static float interpolate(
-      float startValue,
-      float endValue,
-      float startTime,
-      float endTime,
-      float time)
+  public static int interpolate(
+      int startValue,
+      int endValue,
+      int startTime,
+      int endTime,
+      int time)
   {
-    float componentA = startValue * (endTime - time)/(endTime - startTime);
-    float componentB = endValue * (time - startTime)/(endTime - startTime);
+    int componentA = startValue * (endTime - time)/(endTime - startTime);
+    int componentB = endValue * (time - startTime)/(endTime - startTime);
 
     return componentA + componentB;
   }
